fix: rename to generationInteractionLogger.ts
