+++ packages/salesforcedx-vscode-apex/src/oas/generationStrategy/yaml/methodByMethodStrategy.ts
@@ -12,7 +12,7 @@ import { DocumentSymbol } from 'vscode';
 import * as yaml from 'yaml';
 import { SUM_TOKEN_MAX_LIMIT, IMPOSED_FACTOR, PROMPT_TOKEN_MAX_LIMIT } from '..';
 import { nls } from '../../../messages';
-import GenerationInteractionLogger from '../../generationInterationsLogger';
+import GenerationInteractionLogger from '../../generationInteractionLogger';
 import {
   ApexAnnotationDetail,
   ApexClassOASEligibleResponse,
